requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.10
kombu==5.4.2
multidict==6.1.0
paypalrestsdk==1.13.3
pdfkit==1.0.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
qrcode==8.0
requests==2.32.3
schema==0.7.7
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
twilio==9.3.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.14.0
